fix(windows): resolve CI build errors and bump version to 0.2.1
- Add Win32_Security feature for CreateJobObjectW (E0432)
- Rename JOB_OBJECT_LIMIT_KILL_ON_JOB -> JOB_OBJECT_LIMIT_KILL_ON_JOB_CLOSE (windows 0.58)
- Add unsafe impl Send/Sync for ProcessTracker (HANDLE = *mut c_void, E0277)
- Cast htons() result to u32 to match dw_local_port type (E0308/E0277)
- Bump version 0.1.1 -> 0.2.1 in package.json, Cargo.toml, tauri.conf.json, Cargo.lock, package-lock.json
- Update AGENTS.md for three-platform (macOS/Linux/Windows) support

+++ src-tauri/src/process/port_check.rs
@@ -62,16 +62,16 @@ pub fn pids_on_port(port: u16) -> AppResult<Vec<u32>> {
     let row_size = std::mem::size_of::<MIB_TCPROW_OWNER_PID>();
     let header_size = std::mem::size_of::<u32>();
     let mut pids = Vec::new();
-    let target_port = unsafe { htons(port) };
-    for i in 0..entries_count as usize {
-        let offset = header_size + i * row_size;
-        if offset + row_size > buf.len() {
-            break;
-        }
-        let row = unsafe {
-            &*(buf.as_ptr().add(offset) as *const MIB_TCPROW_OWNER_PID)
-        };
-        if row.dw_local_port == target_port && row.dw_state == 2 {
+        let target_port = unsafe { htons(port) } as u32;
+        for i in 0..entries_count as usize {
+            let offset = header_size + i * row_size;
+            if offset + row_size > buf.len() {
+                break;
+            }
+            let row = unsafe {
+                &*(buf.as_ptr().add(offset) as *const MIB_TCPROW_OWNER_PID)
+            };
+            if row.dw_local_port == target_port && row.dw_state == 2 {
             pids.push(row.dw_owning_pid);
         }
     }
